requirements/requirements.txt
bcrypt==4.0.1
fastapi==0.103.0
uvicorn==0.23.2
PyJWT==2.8.0
python-decouple==3.8
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1